requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
requests==2.26.0
redis==3.5.3
python-dotenv==0.19.0
scikit-learn==1.0.2
numpy==1.21.6
pydantic==1.10.7